refactor: update Roles, Provider & Days Off logic

+++ src/controllers/ProviderController.ts
@@ -3,33 +3,36 @@ import providerService from '../services/ProviderService'
 import { isValidObjectId, isValidTime } from '../utils/validators'
 
 class ProviderController {
-  async create(req: Request, res: Response): Promise<Response> {
-    const { userId, workingHours } = req.body
+  async createOrUpdate(req: Request, res: Response): Promise<Response> {
+    const { workingHours, weeklyClosedDays, closedDates } = req.body
+    const userId = req.user?.id
 
     if (!userId || !isValidObjectId(userId))
-      return res.status(400).json({ message: 'UserId is required' })
+      return res.status(400).json({ message: 'UserId é obrigatório' })
+
     if (!workingHours || workingHours.length === 0) {
-      return res.status(400).json({ message: 'Working hours are required' })
+      return res.status(400).json({ message: 'Horário é obrigatório' })
     }
 
-    // Valida se os horários de trabalho estão no formato correto
     const isValidWorkingHours = workingHours.every(
       (wh: { start: string; end: string }) =>
         isValidTime(wh.start) && isValidTime(wh.end)
     )
 
     if (!isValidWorkingHours) {
-      return res.status(400).json({ message: 'Invalid working hours format' })
+      return res.status(400).json({ message: 'Horário inválido.' })
     }
 
     try {
-      const provider = await providerService.createProvider({
+      const provider = await providerService.createOrUpdateProvider({
         userId,
         workingHours,
+        closedDates,
+        weeklyClosedDays,
       })
       return res.status(201).json(provider)
     } catch (error) {
-      return res.status(500).json({ message: 'Error creating provider' })
+      return res.status(500).json({ message: 'Erro criando provider' })
     }
   }
 
@@ -38,61 +41,49 @@ class ProviderController {
       const providers = await providerService.listProviders()
       return res.json(providers)
     } catch (error) {
-      return res.status(500).json({ message: 'Error listing providers' })
+      return res.status(500).json({ message: 'Erro listando providers' })
+    }
+  }
+
+  async getByUserId(req: Request, res: Response): Promise<Response> {
+    const { id } = req.params
+    if (!isValidObjectId(id)) {
+      return res.status(500).json({ message: 'Parâmetros inválidos' })
+    }
+
+    try {
+      const provider = await providerService.getProviderByUserId(id)
+      return res.json(provider)
+    } catch (error) {
+      return res.status(500).json({ message: 'Erro listando provider' })
     }
   }
 
   async getById(req: Request, res: Response): Promise<Response> {
     const { id } = req.params
     if (!isValidObjectId(id)) {
-      return res.status(500).json({ message: 'Invalid parameters' })
+      return res.status(500).json({ message: 'Parâmetros inválidos' })
     }
 
     try {
       const provider = await providerService.getProviderById(id)
       return res.json(provider)
     } catch (error) {
-      return res.status(500).json({ message: 'Error listing provider' })
-    }
-  }
-
-  async update(req: Request, res: Response): Promise<Response> {
-    const { id } = req.params
-    const { workingHours } = req.body
-
-    if (!workingHours) {
-      return res.status(400).json({ message: 'Parâmetros inválidos' })
-    }
-
-    // Se horas de trabalho são fornecidas, valida o formato
-    if (workingHours && workingHours.length > 0) {
-      const isValidWorkingHours = workingHours.every(
-        (wh: { start: string; end: string }) =>
-          isValidTime(wh.start) && isValidTime(wh.end)
-      )
-
-      if (!isValidWorkingHours) {
-        return res.status(400).json({ message: 'Parâmetros inválidos' })
-      }
-    }
-
-    try {
-      const updated = await providerService.updateProvider(id, {
-        workingHours,
-      })
-      return res.json(updated)
-    } catch (error) {
-      return res.status(500).json({ message: 'Error updating provider' })
+      return res.status(500).json({ message: 'Erro listando provider' })
     }
   }
 
   async delete(req: Request, res: Response): Promise<Response> {
     const { id } = req.params
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({ message: 'Id é obrigatório' })
+    }
+
     try {
       await providerService.deleteProvider(id)
       return res.sendStatus(204)
     } catch (error) {
-      return res.status(500).json({ message: 'Error deleting provider' })
+      return res.status(500).json({ message: 'Erro deletando provider' })
     }
   }
 }
